+++ handlers/project.go
@@ -6,20 +6,11 @@ import (
 	"strconv"
 
 	"github.com/gin-gonic/gin"
-	"github.com/realwebdev/Bilal/clockify3/middleware"
 	"github.com/realwebdev/Bilal/clockify3/models"
 )
 
 func CreateProject(h *Handler) gin.HandlerFunc {
 	return func(c *gin.Context) {
-		if err := middleware.AuthenticateToken(c.Request); err != nil {
-			c.JSON(http.StatusUnauthorized, gin.H{
-				"message": "Error in Authorizaition of JWT",
-				"error":   err,
-			})
-			return
-		}
-
 		project := models.Project{}
 		if err := c.BindJSON(&project); err != nil {
 			c.JSON(http.StatusNotFound, gin.H{
@@ -41,14 +32,6 @@ func CreateProject(h *Handler) gin.HandlerFunc {
 
 func GetProjects(h *Handler) gin.HandlerFunc {
 	return func(c *gin.Context) {
-		if err := middleware.AuthenticateToken(c.Request); err != nil {
-			c.JSON(http.StatusUnauthorized, gin.H{
-				"message": "Error in Authorizaition of JWT",
-				"error":   err,
-			})
-			return
-		}
-
 		users, err := h.DB.GetProjects()
 		if err != nil {
 			c.JSON(http.StatusNotFound, gin.H{
@@ -63,14 +46,6 @@ func GetProjects(h *Handler) gin.HandlerFunc {
 
 func UpdateProject(h *Handler) gin.HandlerFunc {
 	return func(c *gin.Context) {
-		if err := middleware.AuthenticateToken(c.Request); err != nil {
-			c.JSON(http.StatusUnauthorized, gin.H{
-				"message": "Error in Authorizaition of JWT",
-				"error":   err,
-			})
-			return
-		}
-
 		uintt, _ := strconv.ParseUint(c.PostForm("id"), 10, 32)
 		id := uint(uintt)
 		update := c.PostForm("update")
@@ -90,14 +65,6 @@ func UpdateProject(h *Handler) gin.HandlerFunc {
 
 func DeleteProject(h *Handler) gin.HandlerFunc {
 	return func(c *gin.Context) {
-		if err := middleware.AuthenticateToken(c.Request); err != nil {
-			c.JSON(http.StatusUnauthorized, gin.H{
-				"message": "Error in Authorizaition of JWT",
-				"error":   err,
-			})
-			return
-		}
-
 		uintt, _ := strconv.ParseUint(c.PostForm("id"), 10, 32)
 		id := uint(uintt)
