+++ routers/router.go
@@ -5,6 +5,7 @@ import (
 
 	"github.com/gin-gonic/gin"
 	"github.com/realwebdev/Bilal/clockify3/handlers"
+	"github.com/realwebdev/Bilal/clockify3/middleware"
 )
 
 func SetupRouter(handler *handlers.Handler) *gin.Engine {
@@ -13,22 +14,27 @@ func SetupRouter(handler *handlers.Handler) *gin.Engine {
 		c.JSON(http.StatusOK, "pong")
 	})
 
-	router.GET("/getusers", handlers.GetUsers(handler))
 	router.POST("/signup", handlers.CreateUser(handler))
 	router.PUT("/signin", handlers.AuthenticateUser(handler))
-	router.DELETE("/deleteuser", handlers.DeleteUser(handler))
 
-	router.GET("/getproject", handlers.GetProjects(handler))
-	router.POST("/createproject", handlers.CreateProject(handler))
-	router.POST("/updateproject", handlers.UpdateProject(handler))
-	router.DELETE("/deleteproject", handlers.DeleteProject(handler))
+	//middleware
+	// authorized := router.Group("/")
+	// authorized.Use(middleware.TokenAuthMiddleware())
+	// {
+	router.GET("/getusers", middleware.TokenAuthMiddleware(), handlers.GetUsers(handler))
+	router.DELETE("/deleteuser", middleware.TokenAuthMiddleware(), handlers.DeleteUser(handler))
 
-	router.POST("/createactivity", handlers.StartActivity(handler))
-	router.POST("/endactivity", handlers.EndActivity(handler))
-	router.POST("/updateactivity", handlers.UpdateActivity(handler))
-	router.DELETE("/deleteactivity", handlers.DeleteActivity(handler))
+	router.GET("/getproject", middleware.TokenAuthMiddleware(), handlers.GetProjects(handler))
+	router.POST("/createproject", middleware.TokenAuthMiddleware(), handlers.CreateProject(handler))
+	router.POST("/updateproject", middleware.TokenAuthMiddleware(), handlers.UpdateProject(handler))
+	router.DELETE("/deleteproject", middleware.TokenAuthMiddleware(), handlers.DeleteProject(handler))
 
-	// router.GET("/isAuthenticated", handlers.DeleteActivity(handler))
+	router.POST("/createactivity", middleware.TokenAuthMiddleware(), handlers.StartActivity(handler))
+	router.POST("/endactivity", middleware.TokenAuthMiddleware(), handlers.EndActivity(handler))
+	router.POST("/updateactivity", middleware.TokenAuthMiddleware(), handlers.UpdateActivity(handler))
+	router.DELETE("/deleteactivity", middleware.TokenAuthMiddleware(), handlers.DeleteActivity(handler))
+
+	// }
 
 	return router
 }
